chore: account for navigator not being ready
